pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "ods_mlops"
authors =[ { name="KernelA"  }]
requires-python = ">=3.10"
version = "0.0.1"
readme = "README.md"
classifiers = [ "Intended Audience :: Developers", "Programming Language :: Python :: 3" ]
dependencies = [
    "pyyaml",
    "hydra-core~=1.3.2",
    "snakemake~=7.32",
    "pulp~=2.7.0",
    "polars~=0.20.19",
    "scikit-learn~=1.4.0",
    "scikit-learn-intelex~=2024.3.0",
    "pandas~=2.2.2",
    "pyarrow~=15.0.1",
    "mplcairo>=0.5",
    "seaborn~=0.13.2",
    "dvc[gdrive]~=3.50",
    "fsspec~=2024.3.0",
]

[project.optional-dependencies]
eda = [
    "plotly~=5.20.0",
]
notebook = [
    "ipykernel~=6.29",
    "ipywidgets~=8.1",
    "nbformat~=5.1.0",
]
[tool.pdm.dev-dependencies]
dev = [
    "ruff~=0.3.0",
    "pytest~=8.1",
    "pre-commit~=3.7",
    "kaggle>=1.6.12",
]

[tool.ruff.lint]
ignore = ["F401", "F403", "E402"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 100

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.pdm.options]
add = ["--no-isolation", "--no-self", "--update-reuse"]
install = ["--no-self"]
lock = ["--update-reuse"]